django/requirements.txt
asgiref==3.11.0
Django==6.0
django-filter==25.2
django-jazzmin==3.0.1
django-vite==3.1.0
djangorestframework==3.16.1
gunicorn==23.0.0
Markdown==3.10
packaging==25.0
prettyprint==0.1.5
psycopg==3.3.2
psycopg-binary==3.3.2
python-decouple==3.8
python-dotenv==1.2.1
sqlparse==0.5.4
whitenoise==6.11.0
